requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==3.2.23
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.4
djangorestframework==3.13.1
gunicorn==21.2.0
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.8.0
whitenoise==6.6.0
